Issue 6189: Added dmSpinlock to dmSDK (#6196)
* Issue 6189: Added dmSpinlock to dmSDK

* typo
